requirements.txt
fastapi
uvicorn
line-bot-sdk
openai
redis